requirements.txt
# Core dependencies
# Add your project dependencies here
# Example:
# requests>=2.31.0
# python-dotenv>=1.0.0

# Core - Multi-agent and LLM
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
openai>=1.0.0
langchain-tavily
pinecone>=5.0.0

# Supporting
python-dotenv>=1.0.0
python-dateutil>=2.8.0
google-search-results>=2.4.0
serpapi>=0.1.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
